Move displayUserBookings to domUpdates
